+++ visualization/app/codeCharta/ui/colorSettingsPanel/metricColorRangeSlider/rangeSlider/rangeSlider.component.ts
@@ -12,7 +12,6 @@ export const sliderWidth = 150
 	template: require("./rangeSlider.component.html")
 })
 export class RangeSliderComponent implements OnChanges {
-	@Input() disabled: boolean
 	@Input() minValue: number
 	@Input() maxValue: number
 	@Input() currentLeftValue: number