+++ visualization/app/codeCharta/ui/colorSettingsPanel/metricColorRangeSlider/rangeSlider/selectors/metricColorRangeSliderColors.selector.ts
@@ -2,16 +2,8 @@ import { createSelector } from "../../../../../state/angular-redux/createSelecto
 import { isDeltaStateSelector } from "../../../../../state/selectors/isDeltaState.selector"
 import { mapColorsSelector } from "../../../../../state/store/appSettings/mapColors/mapColors.selector"
 
-export const metricColorRangeSliderColorsSelector = createSelector([mapColorsSelector, isDeltaStateSelector], (mapColors, isDeltaState) =>
-	isDeltaState
-		? {
-				leftColor: "#DDDDDD",
-				middleColor: "#DDDDDD",
-				rightColor: "#DDDDDD"
-		  }
-		: {
-				leftColor: mapColors.positive,
-				middleColor: mapColors.neutral,
-				rightColor: mapColors.negative
-		  }
-)
+export const metricColorRangeSliderColorsSelector = createSelector([mapColorsSelector, isDeltaStateSelector], mapColors => ({
+	leftColor: mapColors.positive,
+	middleColor: mapColors.neutral,
+	rightColor: mapColors.negative
+}))